Solution for Problem 8: String to Integer (atoi)
This is a Python 3 implementation of the atoi C library
function. The function uses the Python re standard library
regular expression module for this solution implementation.

Consequently, this solution runs faster than only 8% of all
Python 3 submissions. Equivalently, this means this solution
runs slower than 92% of all Python 3 submissions. On the
bright side, at least the code is clean.

It should be noted that the test cases for this problem are
actually pretty pitiful. I did not test whether the input
string results in a number greater than 2^31 - 1, although
the implementation does actually handle this edge case,
in spite of the lacking test coverage.

The LeetCode problem description was not very good, but I
could have done a much better job with the test cases
regardless.