Solution for Problem 7: Reverse Integer
This is the solution for LeetCode problem 7, reversing an integer.
Unlike the previous solution I pushed, the one for problem 8, this
solution implementation ran faster than 84% of all Python 3
submissions, although that was not my explicit intent when I was
formulating this procedure.

The only subtlety in this algorithm is in the division of the
input variable. Naively using the division operator will
implicitly coerce the variable into a floating point variable,
thus rendering the stopping condition fruitless, since the while
loop will terminate if and only if the input variable is zero.

To circumvent this problem, the floordiv operation is used to
truncate any non-integral component of the result, thereby
ensuring that the stop condition of the while loop is fulfilled
eventually.

To this point, I believe we were supposed to not use any numbers
larger than 32 bits at all, rather than simply truncating the
answer down. I didn't actually do this, however, although the
code obviously passed. This should be flagged as a bug in my
eyes.